requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
openai>=1.0.0
PyYAML>=6.0
feedparser>=6.0.10
scikit-learn>=1.3.0
numpy>=1.24.0